replaced all configs by one
