--- hw-02/src/main/java/ru/otus/config/UserInitServiceConfig.java
@@ -1,15 +1,0 @@
-package ru.otus.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import ru.otus.service.InputService;
-import ru.otus.service.UserInitService;
-import ru.otus.service.UserInitServiceImpl;
-
-@Configuration
-public class UserInitServiceConfig {
-    @Bean
-    public UserInitService userInitService(InputService inputService) {
-        return new UserInitServiceImpl(inputService);
-    }
-}
